% SPDX-FileCopyrightText: 2026 Arnaldo Mendes Pires Junior
% SPDX-License-Identifier: MIT
\name{benchmarks}
\alias{sphere}
\alias{rosenbrock}
\alias{rastrigin}
\title{Benchmark objective functions}
\usage{
sphere(x)

rosenbrock(x)

rastrigin(x)
}
\arguments{
  \item{x}{Numeric vector.}
}
\value{Scalar objective value.}
\description{Standard benchmark functions used in optimization tests and examples.}
